DeployML/requirement.txt
asgiref==3.3.1
certifi==2020.12.5
Django==3.1.5
djangorestframework==3.12.2
joblib==1.0.0
numpy==1.19.5
pandas==1.2.0
python-dateutil==2.8.1
pytz==2020.5
scikit-learn==0.24.0
scipy==1.6.0
six==1.15.0
sklearn==0.0
sqlparse==0.4.1
threadpoolctl==2.1.0
wincertstore==0.2
